Refactored Log Progress Popup to different component
